+++ app/components/AddressBlock.js
@@ -1,7 +1,6 @@
 // @flow
 import React, { Component } from 'react';
 import styled, { withTheme } from 'styled-components';
-import { clipboard } from 'electron';
 import { darken } from 'polished';
 import AddCircle from 'material-ui/svg-icons/content/add-circle';
 import CloseIcon from 'material-ui/svg-icons/navigation/close';
@@ -129,21 +128,6 @@ const NoSmartAddressesButton = styled(Button)`
   font-weight: ${({theme: {typo}}) => typo.weights.bold};
   width: 100%;
 `
-const CopyImage = styled.img`
-  margin-left: ${ms(-4)};
-  with: ${ms(1)};
-  height: ${ms(1)};
-  cursor: pointer;
-`
-const CopyContent = styled.span`
-  display: flex;
-  alignItems: center;
-  font-size: ${ms(0)};
-`
-
-const copyToClipboard = text => {
-  clipboard.writeText(text)
-}
 
 const Syncing = styled.div`
   display: ${({ isReady }) =>  isReady ? 'none' : 'flex'};
@@ -239,19 +223,9 @@ class AddressBlock extends Component<Props, State> {
           <TezosAmount
             color={'primary'}
             size={ms(1)}
-            amount={ balance }
+            amount={balance}
             format={2}
             showTooltip
-            content={
-              <CopyContent>
-                {formatedBalance}
-                <CopyImage
-                  src={contentCopy}
-                  style={{fill: theme.colors.gray2}}
-                  onClick={() => copyToClipboard(formatedBalance)}
-                />
-              </CopyContent>
-            }
           />
         </AddressLabel>
         <Address
