+++ app/components/TezosAmount/index.js
@@ -1,9 +1,12 @@
 import * as React from 'react';
 import styled from 'styled-components';
+import { clipboard } from 'electron';
 import TezosIcon from '../TezosIcon';
+import CopyIcon from '../CopyIcon';
 import Tooltip from '../Tooltip';
 import { formatAmount } from "../../utils/currancy";
 import { ms } from '../../styles/helpers';
+import contentCopy from '../../../resources/contentCopy.svg';
 
 type Props = {
   amount: number,
@@ -37,10 +40,37 @@ const Icon = styled(TezosIcon)`
   line-height: 1;
 `;
 
+const CopyImage = styled.img`
+  margin-left: ${ms(-4)};
+  with: ${ms(1)};
+  height: ${ms(1)};
+  cursor: pointer;
+  fill: ${ ({ theme: { colors } }) => colors.gray2 };
+`
+const CopyContent = styled.span`
+  display: flex;
+  alignItems: center;
+  font-size: ${ms(0)};
+`
+const Content = props => {
+  const { formatedBalance } = props
+  return (
+    <CopyContent>
+      {formatedBalance}
+      <CopyIcon text={formatedBalance} color='primary' />
+    </CopyContent>
+  )
+}
+
+const copyToClipboard = text => {
+  clipboard.writeText(text)
+}
+
 const TezosAmount = (props: Props) => {
   const { size, color, amount, iconName, weight, className, showTooltip ,format, content} = props;
+  const formatedBalance = `${formatAmount(amount)}`
   return showTooltip ? (
-    <Tooltip position="right" content={formatAmount(amount)}>
+    <Tooltip position="top" content={<Content formatedBalance={formatedBalance} />}>
       <Amount className={className} color={color} size={size} weight={weight}>
         { format === 6 ? formatAmount(amount) : `~${formatAmount(amount, format)}` }
         {
